requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.4.0
prettytable==3.10.2
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
termcolor==2.4.0
twilio==9.2.3
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
